fix(FEAT-003): cast conviction-derived shares to int, tighten test assertions

+++ agent/main.py
@@ -170,7 +170,7 @@ def cmd_run2(db_path: str = DB_PATH, plan_path: str = PLAN_PATH, output_path: st
                 # FEAT-003: derive shares from conviction + current price
                 conviction = trade.get("conviction", "medium")
                 dollar_amount = position_size_from_conviction(conviction, portfolio["cash"])
-                shares = dollar_amount / current_price
+                shares = int(dollar_amount / current_price)
                 valid, reason = validate_buy(ticker, shares, current_price, portfolio)
                 if valid:
                     execute_buy(ticker, shares, current_price, reasoning, db_path)
